+++ FiNote/www/index.html
@@ -19,7 +19,9 @@
     <script src="js/spin.min.js"></script>
     <script type="text/javascript" src="js/SocialSharing.js"></script>
     <script type="text/javascript" src="js/chartist.min.js"></script>
-    <script src="js/index.js"></script>
+    <script src="js/utility.js"></script>
+	<script src="js/db_method.js"></script>
+	<script src="js/index.js"></script>
 </head>
 
 <body>
